TP4: Added a dual port rom using the quartus megawizzard.
Port B has a 19 bit address bus and a 1 bit address bus. This lets me
read pixel by pixel.

Port A however is a 16 bit address bus, with an 8 bit data bus. This is
because we only have 1.44ms per frame to clear the ram and write the new
data (during the vertical blanking period). Even with a 100MHz clock
(max for SRAM read) it would take > 3ms to set every bit to zero. Since
I have to do this and then write ~12,000 co-ordinates, I've had to make
the data bus 8 bits wide.

In the actual project we have create new data every 3 ticks (from the
cordic) and for each value we have to read the video ram, set the bit
and write it back to the video ram. I've commited a testbench which does
more or less this, as a proof of concept.