TP4: Added a kinetis design studio project.
The K66_FRDM board has a small coprocessor running OpenSDA software.
Which implements a USB CDC serial link and forwards the serial data
to the main processor.

My plan is to have that main processor forward the comms to the FPGA.

This is because my laptop does not have a serial port, and I don't have
a usb to serial converter.