TP4: Adding an error LED to indicate writing to video ram when active.
To avoid double buffering I clear and update the video ram during the
blanking period, IE. when we aren't sending pixel data. If it takes too
long to process all the coordinate data we could overrun the blanking
period. In which case we would start to get errors displayed.

This LED output asures me that this isn't occuring.